pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maestro-flow"
version = "0.1.0"
description = "Multi-agent standardized software delivery workflow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai>=1.54.0",
  "pydantic>=2.9.0",
  "PyYAML>=6.0.2",
  "typer>=0.12.5",
  "rich>=13.8.0",
  "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
test = [
  "pytest>=8.3.0"
]

[project.scripts]
maestro = "maestro_flow.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

